+++ test/test_bases.jl
@@ -38,7 +38,7 @@ comp_b1_b2 = tensor(comp_b1, comp_b2)
 @test comp_b1_b2.shape == [prod(shape1), prod(shape2), prod(shape1), prod(shape1), prod(shape2)]
 @test comp_b1_b2 == CompositeBasis(b1, b2, b1, b1, b2)
 
-@test_throws ArgumentError tensor()
+@test !applicable(tensor)
 @test comp_b2.shape == tensor(b1, comp_b1).shape
 @test comp_b2 == tensor(b1, comp_b1)