+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.4.4 - 2026-08-31
 
-- Use TensorCore's shared `tensor` and `⊗` interface to avoid import conflicts.
+- Use TensorCore's shared `tensor` and `⊗` interface to avoid import conflicts and method piracy.
 
 ## v0.4.3 - 2026-08-21